+++ src/Config/Util/ConfigOptionsValidatorsTrait.php
@@ -5,9 +5,8 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Installer\Config\Util;
 
 use Shlinkio\Shlink\Installer\Exception\InvalidConfigOptionException;
-
 use Shlinkio\Shlink\Installer\Util\Utils;
-use function explode;
+
 use function Functional\map;
 use function is_numeric;
 use function preg_match;