+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [Unreleased]
+### Added
+* [#122](https://github.com/shlinkio/shlink-installer/issues/122) Added support for QR code config options.
+
+### Changed
+* *Nothing*
+
+### Deprecated
+* *Nothing*
+
+### Removed
+* *Nothing*
+
+### Fixed
+* *Nothing*
+
+
 ## [6.1.0] - 2021-08-04
 ### Added
 * [#115](https://github.com/shlinkio/shlink-installer/issues/115) Added support for `append_extra_path` config option.