+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+## [6.2.0] - 2021-10-10
 ### Added
 * [#122](https://github.com/shlinkio/shlink-installer/issues/122) Added support for QR code config options.
 * [#124](https://github.com/shlinkio/shlink-installer/issues/124) Added support for redis sentinels in redis config.
 * [#126](https://github.com/shlinkio/shlink-installer/issues/126) Added support to send orphan visits to webhooks, if any.
+* [#128](https://github.com/shlinkio/shlink-installer/issues/128) Added support for IP-based tracking.
 
 ### Changed
 * *Nothing*